# Remove delta update in Report batch