--- a/include/istio/mixerclient/options.h
+++ b/include/istio/mixerclient/options.h
@@ -45,8 +45,8 @@ struct CheckOptions {
 // Options controlling report batch.
 struct ReportOptions {
   // Default constructor.
-  // Default to batch up to 1000 reports or 1 seconds.
-  ReportOptions() : max_batch_entries(1000), max_batch_time_ms(1000) {}
+  // Default to batch up to 500 reports or 1 seconds.
+  ReportOptions() : max_batch_entries(500), max_batch_time_ms(1000) {}
 
   // Constructor.
   ReportOptions(int max_batch_entries, int max_batch_time_ms)