--- a/src/istio/mixerclient/report_batch_test.cc
+++ b/src/istio/mixerclient/report_batch_test.cc
@@ -23,8 +23,8 @@ using ::google::protobuf::util::error::Code;
 using ::istio::mixer::v1::Attributes;
 using ::istio::mixer::v1::ReportRequest;
 using ::istio::mixer::v1::ReportResponse;
-using ::testing::_;
 using ::testing::Invoke;
+using ::testing::_;
 
 namespace istio {
 namespace mixerclient {
@@ -104,29 +104,6 @@ TEST_F(ReportBatchTest, TestBatchReport) {
 
   batch_->Flush();
   EXPECT_EQ(report_call_count, 4);
-}
-
-TEST_F(ReportBatchTest, TestNoDeltaUpdate) {
-  int report_call_count = 0;
-  EXPECT_CALL(mock_report_transport_, Report(_, _, _))
-      .WillRepeatedly(Invoke([&](const ReportRequest& request,
-                                 ReportResponse* response, DoneFunc on_done) {
-        report_call_count++;
-        on_done(Status::OK);
-      }));
-
-  Attributes report;
-  utils::AttributesBuilder(&report).AddString("key", "value");
-  batch_->Report(report);
-  EXPECT_EQ(report_call_count, 0);
-
-  // Erase a key, so delta update fail to push the batched result.
-  report.mutable_attributes()->erase("key");
-  batch_->Report(report);
-  EXPECT_EQ(report_call_count, 1);
-
-  batch_->Flush();
-  EXPECT_EQ(report_call_count, 2);
 }
 
 TEST_F(ReportBatchTest, TestBatchReportWithTimeout) {