… src/istio/mixerclient/delta_update.cc
… src/istio/mixerclient/delta_update.h
… src/istio/mixerclient/delta_update_test.cc