Signed-off-by: Wayne Zhang <qiwzhang@google.com>

**What this PR does / why we need it**:

DeltaUpdate is comparing each attribute with the last Report in the Batch.  It is very slow, use a lot of cpu.  It compressed some bytes, but bandwidth is not the issue for Istio, cpu usage cause latency.
Hence remove the delta update.

Also reduce the batch size since the byte could not be compressed much.  Sending a huge batch report which may cause some requests to get delayed more (if these requests are in the queue). It will cause 99 percentitle latency to increase.

**Special notes for your reviewer**:

**Release note**:
```release-note
None
```