@istio-testing 2018-09-13
@qiwzhang: The following tests **failed**, say `/retest` to rerun them all:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
prow/proxy-presubmit.sh | e70d018337da094dd956e7e78e42ba068464a57c | [link](https://k8s-gubernator.appspot.com/build/istio-prow/pull/istio_proxy/1970/proxy-presubmit/2203/) | `/test proxy-presubmit`
prow/proxy-presubmit-asan.sh | e70d018337da094dd956e7e78e42ba068464a57c | [link](https://k8s-gubernator.appspot.com/build/istio-prow/pull/istio_proxy/1970/proxy-presubmit-asan/1459/) | `/test proxy-presubmit-asan`



<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->