@mandarjog 2018-09-13
@qiwzhang can you put before and after latency numbers here?

Do you want to keep it as an option until we are done with the performance work?