@qiwzhang 2018-09-13
Hmm,  the test result shows:  for fortio-c=10, envoy-concurrency=1, report-only.

1) 50%tile: is faster from 2.2ms to 1.7ms
2) 99%tile: is slower from 5.4ms to 15.6ms if batch_numbers only reduced to 500 from 1000.  if reduces to 100:  from 5.4ms to 4.7ms

If max_number can be reduced to 100.  I think we can keep this change.   We like to have .5ms latency decrease for average request.  99% tile is good too.

I think qps to maxer will NOT be increased by 10 times.  Most report batches are flushed out by 1 second timer instead of max_numbers.